requirements.txt
numpy
pandas
ta
gymnasium
stable-baselines3[extra]
optuna
matplotlib
scikit-learn
pyyaml
quantstats
shimmy>=2.0
